Fix build error with MSVC 2008 and 2010
git-svn-id: https://svn.osgeo.org/gdal/trunk@34237 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/alg/gdal_rpc.cpp
@@ -1181,7 +1181,8 @@ static bool GDALRPCExtractDEMWindow( GDALRPCTransformInfo *psTransform,
         int nRadius = psTransform->nBufferMaxRadius;
         if( psTransform->nDEMExtractions < psTransform->nBufferMaxRadius * psTransform->nBufferMaxRadius )
         {
-            nRadius = static_cast<int> (sqrt(psTransform->nDEMExtractions) );
+            nRadius = static_cast<int> (
+                  sqrt(static_cast<double>(psTransform->nDEMExtractions)) );
             CPLAssert( nRadius <= psTransform->nBufferMaxRadius );
             if( nRadius < nWidth )
                 nRadius = nWidth;
